requirements.txt
fastapi
uvicorn
feedparser
transformers
torch
supabase
python-dotenv
requests
httpx
pydantic